Memory leak cleanup (#38)
Cleanup of SimpleAV to remove everything correctly. Convert SimpleAV and BackgroundAudio to use pool of Hls instances. This has added advantage of also abstracting Hls.js so it can be replaced with dash.js if needed at a later date.

This code fixes a good proportion of leakage on Linux Chromium and almost all leakage on Linux Firefox. When merged I will test on Windows Chrome and Firefox. The remaining leakage in Linux Chromium still happens when Hls.js is replaced with dash.js so I suspect it's a Linux Chrome specific issue (MSE buffer not being cleaned by browser).